+++ packages/vite/src/node/ssr/ssrTransform.ts
@@ -39,6 +39,7 @@ export async function ssrTransform(
   let uid = 0
   const deps = new Set<string>()
   const idToImportMap = new Map<string, string>()
+  const declaredConst = new Set<string>()
 
   function defineImport(node: Node, source: string) {
     deps.add(source)
@@ -154,6 +155,16 @@ export async function ssrTransform(
         ) {
           s.appendLeft(id.end, `: ${binding}`)
         }
+      } else if (
+        parent.type === 'ClassDeclaration' &&
+        id === parent.superClass
+      ) {
+        if (!declaredConst.has(id.name)) {
+          declaredConst.add(id.name)
+          // locate the top-most node containing the class declaration
+          const topNode = parentStack[1]
+          s.prependRight(topNode.start, `const ${id.name} = ${binding};\n`)
+        }
       } else {
         s.overwrite(id.start, id.end, binding)
       }
@@ -212,11 +223,11 @@ function walk(
 
   ;(eswalk as any)(root, {
     enter(node: Node, parent: Node | null) {
-      parent && parentStack.push(parent)
-
       if (node.type === 'ImportDeclaration') {
         return this.skip()
       }
+
+      parent && parentStack.push(parent)
 
       if (node.type === 'MetaProperty' && node.meta.name === 'import') {
         onImportMeta(node)
@@ -305,6 +316,11 @@ function isRefIdentifier(id: Identifier, parent: _Node, parentStack: _Node[]) {
     if (parent.params.includes(id)) {
       return false
     }
+  }
+
+  // class method name
+  if (parent.type === 'MethodDefinition') {
+    return false
   }
 
   // property key